+++ requests_cache/backends/sqlite.py
@@ -2,14 +2,15 @@ import sqlite3
 import threading
 from contextlib import contextmanager
 from logging import getLogger
-from os import makedirs
-from os.path import abspath, basename, dirname, expanduser, isabs, join
+from os import makedirs, unlink
+from os.path import abspath, basename, dirname, expanduser, isabs, isfile, join
 from pathlib import Path
 from tempfile import gettempdir
 from typing import Collection, Iterable, Iterator, List, Tuple, Type, Union
 
 from . import BaseCache, BaseStorage, get_valid_kwargs
 
+SQLITE_MAX_VARIABLE_NUMBER = 999
 logger = getLogger(__name__)
 
 
@@ -47,6 +48,15 @@ class DbCache(BaseCache):
         self.redirects.bulk_delete(values=keys)
         self.redirects.vacuum()
 
+    def clear(self):
+        """Clear the cache by deleting the cache file and re-initializing. This is done to allow
+        clear() to succeed even if the file is corrupted.
+        """
+        if isfile(self.responses.db_path):
+            unlink(self.responses.db_path)
+        self.responses.init_db()
+        self.redirects.init_db()
+
 
 class DbDict(BaseStorage):
     """A dictionary-like interface for SQLite.
@@ -77,19 +87,23 @@ class DbDict(BaseStorage):
         self.fast_save = fast_save
         self.table_name = table_name
 
+        self._lock = threading.RLock()
         self._can_commit = True
         self._local_context = threading.local()
-        with sqlite3.connect(self.db_path, **self.connection_kwargs) as con:
-            self._create_table(con)
+        self.init_db()
 
-    # Initial CREATE TABLE must happen in shared connection; subsequent queries will use thread-local connections
-    def _create_table(self, connection):
-        connection.execute(f'CREATE TABLE IF NOT EXISTS {self.table_name} (key PRIMARY KEY, value)')
+    def init_db(self):
+        """Initialize the database, if it hasn't already been.
+        This must be done in shared connection, but all subsequent queries can use thread-local connections.
+        """
+        self.close()
+        with self._lock, sqlite3.connect(self.db_path, **self.connection_kwargs) as con:
+            con.execute(f'CREATE TABLE IF NOT EXISTS {self.table_name} (key PRIMARY KEY, value)')
 
     @contextmanager
     def connection(self, commit=False) -> Iterator[sqlite3.Connection]:
         """Get a thread-local database connection"""
-        if not hasattr(self._local_context, 'con'):
+        if not getattr(self._local_context, 'con', None):
             logger.debug(f'Opening connection to {self.db_path}:{self.table_name}')
             self._local_context.con = sqlite3.connect(self.db_path, **self.connection_kwargs)
             if self.fast_save:
@@ -97,6 +111,12 @@ class DbDict(BaseStorage):
         yield self._local_context.con
         if commit and self._can_commit:
             self._local_context.con.commit()
+
+    def close(self):
+        """Close any active connections"""
+        if getattr(self._local_context, 'con', None):
+            self._local_context.con.close()
+            self._local_context.con = None
 
     @contextmanager
     def bulk_commit(self):
@@ -119,9 +139,7 @@ class DbDict(BaseStorage):
             self._can_commit = True
 
     def __del__(self):
-        """Close any active connections"""
-        if hasattr(self._local_context, 'con'):
-            self._local_context.con.close()
+        self.close()
 
     def __delitem__(self, key):
         with self.connection(commit=True) as con:
@@ -154,24 +172,25 @@ class DbDict(BaseStorage):
             return con.execute(f'SELECT COUNT(key) FROM  {self.table_name}').fetchone()[0]
 
     def bulk_delete(self, keys=None, values=None):
-        """Delete multiple keys from the cache. Does not raise errors for missing keys.
+        """Delete multiple keys from the cache, without raising errors for any missing keys.
         Also supports deleting by value.
         """
         if not keys and not values:
             return
 
         column = 'key' if keys else 'value'
-        marks, args = _format_sequence(keys or values)
-        statement = f'DELETE FROM {self.table_name} WHERE {column} IN ({marks})'
-
         with self.connection(commit=True) as con:
-            con.execute(statement, args)
+            # Split into small enough chunks for SQLite to handle
+            for chunk in chunkify(keys or values):
+                marks, args = _format_sequence(chunk)
+                statement = f'DELETE FROM {self.table_name} WHERE {column} IN ({marks})'
+                con.execute(statement, args)
 
     def clear(self):
         with self.connection(commit=True) as con:
             con.execute(f'DROP TABLE IF EXISTS {self.table_name}')
-            self._create_table(con)
-            con.execute('VACUUM')
+        self.init_db()
+        self.vacuum()
 
     def vacuum(self):
         with self.connection(commit=True) as con:
@@ -189,6 +208,13 @@ class DbPickleDict(DbDict):
 
     def __getitem__(self, key):
         return self.serializer.loads(super().__getitem__(key))
+
+
+def chunkify(iterable: Iterable, max_size=SQLITE_MAX_VARIABLE_NUMBER) -> Iterator[List]:
+    """Split an iterable into chunks of a max size"""
+    iterable = list(iterable)
+    for index in range(0, len(iterable), max_size):
+        yield iterable[index : index + max_size]
 
 
 def _format_sequence(values: Collection) -> Tuple[str, List]: